+++ Locales/Locale-enUS.lua
@@ -34,7 +34,7 @@ L.InvalidTrackerExclusion = "Cannot exclude parent objective."
 L.ObjectiveIsExcluded = "Objective is already being excluded."
 
 L.InvalidCraftSkillID = "Invalid tradeskill name."
-L.MissingCraftRecipeName = "Please specify a tradeskill recipe name"
+L.MissingCraftRecipeName = "Please specify a Recipe String"
 L.UnknownRecipe = "You do not know the recipe: %s"
 
 L.invalidSyntax = function(err) return "Syntax error: "..err end
@@ -139,7 +139,7 @@ L["Custom"] = true
 L["Delete Tracker"] = true
 L["Delete Objective"] = true
 L["Action"] = true
-L["Display Reference"] = true
+L["Action"] = true
 L["Duplicate"] = true
 L["Duplicate Objective"] = true
 L["Enabled"] = true
@@ -171,9 +171,9 @@ L["TRUE"] = true
 L["Type"] = true
 
 
-L.DisplayReferenceDescription = [[Display References allow you to set which item/currency you want to use for automatic objective information. This includes the icon chosen when using "Automatic Icon" and the item associated with a button's "use" attribute. However, when set to a macrotext, the icon will be unaffected.
+L.DisplayReferenceDescription = [[Actions allow you to set which item/currency you want to use for automatic objective information. This includes the icon chosen when using "Automatic Icon" and the item associated with a button's "use" attribute. However, when set to a macrotext, the icon will be unaffected.
 
-Farming Bar provides a /craft command that you can use in your macrotexts. Simply use "/craft tradeskill recipe name". For example:
+Farming Bar provides a /craft command that you can use in your macrotexts. Simply use "/craft Recipe String". For example:
 
 /craft Enchanting Eternal Bounds
 /use Enchanting Vellum]]
@@ -199,7 +199,7 @@ L["Objective Complete"] = true
 L["Counts For"] = true
 L["Currency ID/Link"] = true
 L["Item ID/Name/Link"] = true
-L["Tradeskill Recipe Name"] = true
+L["Recipe String"] = true
 
 ------------------------------------------------------------